+++ CHANGELOG.md
@@ -1,6 +1,17 @@
-# Change Log
+# Pyrsia VS Code Extension Changelog
+
 ## [Unreleased]
-# JFrog MS Teams App Changelog
+Not available in the VS Code store yet.
 
 ## [0.0.1] - Apr 25th, 2022
-* Initial API, views and the basic Pyrsia node support.
+* Initial release. Features:
+  * Views: 
+    * Pyrsia node connection status.
+    * Integrations (Docker support).
+    * Help.
+  * New commands: 
+    * Update Pyrsia node configuration.
+    * Update the docker config.
+    * Add docker image to Pyrsia (request build).
+    * Open the transparency log (docker images).
+    * Request to replace the local docker images with Pyrsia images.